Migrated Files From FTP Server
I migrated most of the files from the FTP Server
hsmscreativitycenter.org/DaMahdi03, as it is currently used as a website
host and I wanted to save/backup all of my files. I also took advantage of
GitHub Pages because it is a free way to spread your code as a website
and it has no hassle involved.

This Repository contains files that are mainly used in Static Website
Development (HTML, CSS, JavaScript, JSON, XML, PHP, MYSQL). I surely
hope to one day make most of those into Dynamic Websites with my new
understanding of Python and maybe even learn Apache to make it into a website
with full control.

In addition to all of these files, I'm creating new websites that are
already in Development. They will not be revealed yet because I don't
want anybody else to get the credit. They too will be Static Websites,
but because they will have a common theme, I think I will design and
build a Python Console Application that can help the repitition of
creating these Static Webpages. If anyone can help me, please feel free
to reach out to me!